* object.c (arrayins2): use a cast to "gsize".

+++ src/object.c
@@ -635,7 +635,7 @@ arrayins2(struct narray *array, const char *val, unsigned int idx)
   if (array->num == array->size) {
     int size;
     size = array->size+ALLOCSIZE;
-    data = g_realloc(array->data,array->base*size);
+    data = g_realloc(array->data,(gsize) array->base*size);
     if (data == NULL) {
       return NULL;
     }
